added printDevices method, created test.class to test methods, changed float id of Products into int

+++ src/products/Smartphone.java
@@ -3,7 +3,7 @@ package products;
 public class Smartphone extends Product {
 
     public Smartphone(Devices deviceType, String brand, String model, String description, float display, float storage,
-            float salePrice, float purchasePrice, float id) {
+                      float salePrice, float purchasePrice, int id) {
         super(deviceType, brand, model, description, display, storage, salePrice, purchasePrice, id);
     }
 }
